refactor(website): extract the message article from the inbox preview
